authentication and user creation now functional

+++ app/controllers/posts_controller.rb
@@ -10,35 +10,41 @@ class PostsController < ApplicationController
   end
   
   def create
-    post = Post.new(create_post_params)
+    if current_user.author?
+      post = Post.new(create_post_params)
 
-    if post.save
-      render json: post
-    else
-      render json: { status: :failed, errors: post.errors }, status: :bad_request
+      if post.save
+        render json: post
+      else
+        render json: { status: :failed, errors: post.errors }, status: :bad_request
+      end
     end
   end
 
   def update
-    post = Post.find(params[:id])
-    post.assign_attributes(update_post_params)
-    if post.save
-      render json: post
-    else
-      render json: { status: :failed, errors: post.errors }, status: :bad_request
+    if current_user.author?
+      post = Post.find(params[:id])
+      post.assign_attributes(update_post_params)
+      if post.save
+        render json: post
+      else
+        render json: { status: :failed, errors: post.errors }, status: :bad_request
+      end
     end
   end
 
   def destroy
-    post = Post.find(params[:id])
-    post.destroy
-    head :ok
+    if current_user.author?
+      post = Post.find(params[:id])
+      post.destroy
+      head :ok
+    end
   end
 
   private
 
   def create_post_params
-    params.require(:post).permit(:title, :subtitle, :content)
+    params.require(:post).permit(:title, :subtitle, :content, :user_id)
   end
 
   def update_post_params
